Rename child.js to processChild.js
