+++ packages/jest-worker/src/workers/processChild.js
